fix build (#450)
