+++ git-repository/src/remote/connection/fetch/update_refs.rs
@@ -55,7 +55,7 @@ pub mod update {
     }
 }
 
-/// Information about the update of a single reference, corresponding the respective entry in [`RefMap::mapping`].
+/// Information about the update of a single reference, corresponding the respective entry in [`RefMap::mappings`][crate::remote::fetch::RefMap::mappings].
 #[derive(Debug, Clone, Copy)]
 pub struct Update {
     /// The way the update was performed.
